mod_7_apiKey/cloud_function/requirements.txt
# Function dependencies, for example:
# package>=version

functions-framework==3.*
google-cloud-logging==3.0.0
openai==0.27.2
oauth2client==4.1.3
google-cloud-ndb==2.1.1

